% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{acl_plot}
\alias{acl_plot}
\title{FUNCTION_TITLE}
\usage{
acl_plot(samples, lag_max = 30, latex = FALSE)
}
\arguments{
\item{samples}{PARAM_DESCRIPTION}

\item{lag_max}{PARAM_DESCRIPTION, Default: 30}

\item{latex}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
