% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{density_plot}
\alias{density_plot}
\title{Plot kernel density estimate for values of a single Markov Chain}
\usage{
density_plot(samples, log = FALSE, robust = FALSE, latex = FALSE)
}
\arguments{
\item{samples}{Numeric vector, matrix with a single column or list
containing such a vector or matrix as the only element.}

\item{log}{Logical. If TRUE, the x - axis is transformed to the
logarithmic scale. \cr
Default: FALSE}

\item{robust}{Logical. If TRUE, the first and last percentile of the samples'
distribution are omitted to avoid a strong influence of
outliers on the x - axis scale. \cr
Default: FALSE}

\item{latex}{Logical. If TRUE, mathematical symbols such as greek letters
in the plot title with subscripts and superscripts are properly
rendered.
This option requires the column of the input matrix to be
labeled accordingly (as in the output of the \code{gibbs_sampler()}
function). \cr
Default: FALSE}
}
\value{
Plot object of the class "ggplot".
}
\description{
The \code{density_plot()} function creates a kernel density plot of
Markov Chain samples with the sampled values mapped to the
x - axis.
The individual data points are displayed by gray points on the
x - axis to get a more detailed overview of the distribution.
}
\examples{
fit <- lslm(
  location = y ~ x1 + x2 + z1 + z2, scale = ~ z1 + z2,
  data = toy_data, light = FALSE
) \%>\%
  gibbs_sampler(num_sim = 1000)

# list of 4 matrices with 1000 rows each
samples <- fit$mcmc_ridge$coefficient_sampling_matrices

# extract first chain of the scale parameter
density_plot(samples$scale[, 1], robust = TRUE)

# log argument often useful for strictly positive variance parameters
density_plot(samples$scale_prior, log = TRUE, latex = TRUE)

}
