% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{FUNCTION_TITLE}
\usage{
diagnostic_plots(
  samples,
  lag_max = 30,
  log = FALSE,
  robust = FALSE,
  latex = FALSE
)
}
\arguments{
\item{samples}{PARAM_DESCRIPTION}

\item{lag_max}{PARAM_DESCRIPTION, Default: 30}

\item{robust}{PARAM_DESCRIPTION, Default: FALSE}

\item{latex}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
