% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{FUNCTION_TITLE}
\usage{
gibbs_sampler(
  m = NULL,
  X = NULL,
  Z = NULL,
  y = NULL,
  beta_start = NULL,
  gamma_start = NULL,
  tau_start = 1,
  xi_start = 1,
  a_tau = 1,
  b_tau = 3,
  a_xi = 1,
  b_xi = 3,
  prop_var = 3,
  num_sim = 1000
)
}
\arguments{
\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{X}{PARAM_DESCRIPTION, Default: NULL}

\item{Z}{PARAM_DESCRIPTION, Default: NULL}

\item{y}{PARAM_DESCRIPTION, Default: NULL}

\item{beta_start}{PARAM_DESCRIPTION, Default: NULL}

\item{gamma_start}{PARAM_DESCRIPTION, Default: NULL}

\item{tau_start}{PARAM_DESCRIPTION, Default: 1}

\item{xi_start}{PARAM_DESCRIPTION, Default: 1}

\item{a_tau}{PARAM_DESCRIPTION, Default: 1}

\item{b_tau}{PARAM_DESCRIPTION, Default: 3}

\item{a_xi}{PARAM_DESCRIPTION, Default: 1}

\item{b_xi}{PARAM_DESCRIPTION, Default: 3}

\item{prop_var}{PARAM_DESCRIPTION, Default: 3}

\item{num_sim}{PARAM_DESCRIPTION, Default: 1000}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}

}
