% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thinning}
\alias{thinning}
\title{Remove samples of a Markov Chain by a constant frequency \code{freq}}
\usage{
thinning(samples, freq)
}
\arguments{
\item{samples}{Numeric vector, matrix or list of matrices containing the
samples of the MCMC sampler.}

\item{freq}{Positive integer which controls the degree to which the
Markov-Chain is thinned out, e.g. \code{freq = 10} only keeps every
10th sample.}
}
\value{
Same data type as the provided input, i.e. a numeric vector, matrix or list
of matrices with thinned out elements.
}
\description{
\code{thinning()} takes a numeric vector, matrix or list of matrices as input and
returns a modified copy of the input object with only a fraction of the
original entries.
The function is primarily used in the context of Markov-Chain Monte-Carlo
sampling to obtain uncorrelated samples of the posterior distribution.
}
