% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{time_plot}
\alias{time_plot}
\title{FUNCTION_TITLE}
\usage{
time_plot(samples, log = FALSE, robust = FALSE, latex = FALSE)
}
\arguments{
\item{samples}{PARAM_DESCRIPTION}

\item{log}{PARAM_DESCRIPTION}

\item{robust}{PARAM_DESCRIPTION, Default: FALSE}

\item{latex}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
