% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Simulation data}
\format{
A data frame with 50 rows and 5 columns:
\describe{
\item{y}{Numeric vector representing observations}
\item{x1, x2}{Regressors for the location parameter}
\item{z1, z2}{Regressors for the location as well as the scale parameter}
}
}
\usage{
toy_data
}
\description{
Simulated data set constructed for unit testing and illustration package
functionality in documentation examples as well as vignettes.
}
\keyword{datasets}
