% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{acl_plot}
\alias{acl_plot}
\title{Plot autocorrelations of a single Markov Chain}
\usage{
acl_plot(samples, lag_max = 30, latex = FALSE)
}
\arguments{
\item{samples}{Numeric vector, matrix with a single column or list
containing such a vector or matrix as the only element.}

\item{lag_max}{Positive integer representing the maximum number of lags
that are shown on the x - axis. \cr
Default: 30}

\item{latex}{Logical. If TRUE, mathematical symbols such as greek letters
in the plot title with subscripts and superscripts are properly
rendered.
This option requires the column of the input matrix to be
labeled accordingly (as in the output of the \code{gibbs_sampler()}
function). \cr
Default: FALSE}
}
\value{
Plot object of the class "ggplot".
}
\description{
The \code{acl_plot()} function creates a graphical display of
autocorrelations for various lags of a single Markov Chain.
The lags are mapped to the x - axis in increasing order with the
corresponding autocorrelation values mapped to the y - axis. \cr
This type of plot can be used to analyze the dependence of the
collected samples i.e. from a posterior distribution.
}
\examples{
fit <- lslm(
  location = y ~ x1 + x2 + z1 + z2, scale = ~ z1 + z2,
  data = toy_data, light = FALSE
) \%>\%
  gibbs_sampler(num_sim = 1000)

# list of 4 matrices with 1000 rows each
samples <- fit$mcmc_ridge$sampling_matrices

# uses default lag_max = 30
acl_plot(samples$scale_prior, latex = TRUE)

# value of lag_max should be adapted depending on the correlation structure
acl_plot(samples$scale[, 1, drop = FALSE], lag_max = 100, latex = TRUE)

}
