% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_plot.R
\name{mult_plot}
\alias{mult_plot}
\title{FUNCTION_TITLE}
\usage{
mult_plot(
  samples,
  type = c("time", "density", "both"),
  log = FALSE,
  robust = FALSE,
  free_scale = FALSE,
  latex = FALSE
)
}
\arguments{
\item{type}{PARAM_DESCRIPTION, Default: c("time", "density", "both")}

\item{robust}{PARAM_DESCRIPTION, Default: FALSE}

\item{free_scale}{PARAM_DESCRIPTION, Default: FALSE}

\item{latex}{PARAM_DESCRIPTION, Default: FALSE}

\item{samples_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
