% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_complete.R
\name{summary_complete}
\alias{summary_complete}
\title{FUNCTION_TITLE}
\usage{
summary_complete(samples, include_plot = FALSE)
}
\arguments{
\item{samples}{PARAM_DESCRIPTION}

\item{include_plot}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
